Day 20 Part 1 - Finally done!
This was a long time coming, and my solution here looks overengineered
and underoptimized.

Some fun timing notes btw: as written here (with a DFS inside the search
function) this takes about 440 ms to run a release build - but 12.3 s
on a debug build! And if I switch the search to a BFS by popping from
the front of the queue instead that debug vs release gap grows even wider,
to 800 ms vs 23.3 s.